+++ TroveTools.NET/View/MyModsView.xaml.cs
@@ -53,5 +53,10 @@ namespace TroveTools.NET.View
                 log.Error("Error adding mods", ex);
             }
         }
+
+        private void ToolBar_Loaded(object sender, RoutedEventArgs e)
+        {
+            MainWindowView.HideToolbarOverflow(sender as ToolBar);
+        }
     }
 }